Created validator.js for validating user info
