+++ routes/instagram.js
@@ -3,7 +3,9 @@ const router = express.Router();
 
 const instagram = require('../mock-database');
 
+const validateInfo = require('../helpers/validator.js')
 const helpers = require('../helpers');
+
 
 // const findUserByName = name => instagram.find(currentUser => currentUser.name === name);
 
@@ -28,7 +30,7 @@ router.put('/:id', (req, res) => {
 	if (!putUser) return res.status(404).send('Unable to edit user.  The ID was not found.');
 	const {
 		error
-	} = helpers.validateInfo(req.body);
+	} = validateInfo(req.body);
 	if (error) return res.status(400).send(error.details[0].message);
 
 	putUser.name = req.body.name;
@@ -40,7 +42,7 @@ router.put('/:id', (req, res) => {
 router.post('/', (req, res) => {
 	const {
 		error
-	} = helpers.validateInfo(req.body); // result.error
+	} = validateInfo(req.body); // result.error
 	if (error) return res.status(400).send(error.details[0].message);
 
 	const newUser = {
